Initial commit of cGAN model
Current task is to ensure that this model runs as a docker image before making changes to the model itself.
Next steps are to ensure environment is set up correctly and configure dockerfile to download the project data from an online source and test if the code works. Then, we can develop the code towards the gap-filling usage.